Feat: otimização de código, centralização das conexões, commands. Usar interface é melhor que usar implementação.
